+++ frontend/src/app/models/page-results.ts
@@ -1,12 +1,10 @@
-import { BehaviorSubject, Observable, combineLatest, from, of } from 'rxjs';
+import { Observable, combineLatest, from, of } from 'rxjs';
 import { QueryModel } from './query';
 import { map } from 'rxjs/operators';
-import { FoundDocument } from './found-document';
 import { SearchService } from '../services';
 import { SearchResults } from './search-results';
-import { CorpusField } from './corpus';
-import * as _ from 'lodash';
 import { Results } from './results';
+import { DocumentPage } from './document-page';
 
 const RESULTS_PER_PAGE = 20;
 
@@ -15,45 +13,6 @@ export interface PageResultsParameters {
     size: number;
 }
 
-export class DocumentPage {
-    focus$ = new BehaviorSubject<FoundDocument>(undefined);
-
-    constructor(
-        public documents: FoundDocument[],
-        public total: number,
-        public fields?: CorpusField[]
-    ) {
-        this.documents.forEach((d, i) => d.position = i + 1);
-    }
-
-    focus(document: FoundDocument) {
-        this.focus$.next(document);
-    }
-
-    focusNext() {
-        this.focusShift(1);
-    }
-
-    focusPrevious() {
-        this.focusShift(-1);
-    }
-
-    blur() {
-        this.focus$.next(undefined);
-    }
-
-    private focusShift(shift: number) {
-        const document = this.focus$.value;
-        if (document) {
-            this.focusPosition(document.position + shift);
-        }
-    }
-
-    private focusPosition(position: number) {
-        const index = _.clamp(position - 1, 0, this.documents.length - 1);
-        this.focus(this.documents[index]);
-    }
-}
 const parseResults = (results: SearchResults): DocumentPage =>
     new DocumentPage(results.documents, results.total.value, results.fields);
